ml/requirements.txt
Flask==3.0.3
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
SQLAlchemy==2.0.29
mysql-connector-python==8.3.0
Flask-Cors==4.0.0

gunicorn
